Cache plot contents in an offscreen surface
When displaying a plot in the buffer, cache the drawn version
of the plot in an offscreen cairo surface so that subsequent
exposes are much faster. This is needed because drawing a
small portion of the plot is going to involve the same python
code and cairo calls as drawing the whole plot, and (at best)
we can only save the actual pixels drawing. With a plot with
large numbers of points, the first steps in the pipeline are
already too slow.

(Suggestion from Jon Kuhn)