Add tests for 'statement-executing' signal ordering and emission
While it isn't guaranteed that any given statement will trigger a
'statement-executing' signal, a 'statement-executing' signal for a statement
should never occur after a 'statement-complete' signal for the same
statement. In addition, any sufficiently long-running statement (in testing,
defined as anything running long enough to be interrupted) would be
expected to trigger a 'statement-executing' signal.