Comprehensively use 'is None' and 'is not None'
Use 'is None' and 'is not None' everywhere; this was recommended
by Jonathan Griffitts in http://www.reinteract.org/trac/ticket/71.
Going ahead and doing it now was inspired by the xmlrpc horror
in http://www.reinteract.org/trac/ticket/77.

No Reinteract object will ever be badly behaved enough to try and
make a remote procedure call when its __eq__() method is called,
but it's easier to make a blanket search and replace than audit
everywhere for user-supplied objects.